Add Scene Delegate & fix to use RestClient extension for iOS13
