+++ iOS13OfflineNativeSwiftTemplate/iOS13OfflineNativeSwiftTemplate/SceneDelegate.swift
@@ -1,0 +1,112 @@
+/*
+Copyright (c) 2019-present, salesforce.com, inc. All rights reserved.
+
+Redistribution and use of this software in source and binary forms, with or without modification,
+are permitted provided that the following conditions are met:
+* Redistributions of source code must retain the above copyright notice, this list of conditions
+and the following disclaimer.
+* Redistributions in binary form must reproduce the above copyright notice, this list of
+conditions and the following disclaimer in the documentation and/or other materials provided
+with the distribution.
+* Neither the name of salesforce.com, inc. nor the names of its contributors may be used to
+endorse or promote products derived from this software without specific prior written
+permission of salesforce.com, inc.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR
+IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
+FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY
+WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+
+import UIKit
+import SwiftUI
+import MobileSync
+import SalesforceSDKCore
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+    
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        self.window = UIWindow(frame: windowScene.coordinateSpace.bounds)
+        self.window?.windowScene = windowScene
+       
+        AuthHelper.registerBlock(forCurrentUserChangeNotifications: {
+           self.resetViewState {
+               self.setupRootViewController()
+           }
+        })
+    }
+
+    func sceneDidDisconnect(_ scene: UIScene) {
+        // Called as the scene is being released by the system.
+        // This occurs shortly after the scene enters the background, or when its session is discarded.
+        // Release any resources associated with this scene that can be re-created the next time the scene connects.
+        // The scene may re-connect later, as its session was not neccessarily discarded (see `application:didDiscardSceneSessions` instead).
+    }
+
+    func sceneDidBecomeActive(_ scene: UIScene) {
+        // Called when the scene has moved from an inactive state to an active state.
+        // Use this method to restart any tasks that were paused (or not yet started) when the scene was inactive.
+        self.initializeAppViewState()
+        AuthHelper.loginIfRequired {
+           // Setup store based on config userstore.json
+           MobileSyncSDKManager.shared.setupUserStoreFromDefaultConfig()
+           // Setup syncs based on config usersyncs.json
+           MobileSyncSDKManager.shared.setupUserSyncsFromDefaultConfig()
+           self.setupRootViewController()
+       }
+    }
+
+    func sceneWillResignActive(_ scene: UIScene) {
+        // Called when the scene will move from an active state to an inactive state.
+        // This may occur due to temporary interruptions (ex. an incoming phone call).
+    }
+
+    func sceneWillEnterForeground(_ scene: UIScene) {
+        // Called as the scene transitions from the background to the foreground.
+        // Use this method to undo the changes made on entering the background.
+    }
+
+    func sceneDidEnterBackground(_ scene: UIScene) {
+        // Called as the scene transitions from the foreground to the background.
+        // Use this method to save data, release shared resources, and store enough scene-specific state information
+        // to restore the scene back to its current state.
+    }
+    
+    // MARK: - Private methods
+   func initializeAppViewState() {
+       if (!Thread.isMainThread) {
+           DispatchQueue.main.async {
+               self.initializeAppViewState()
+           }
+           return
+       }
+       
+       self.window?.rootViewController = InitialViewController(nibName: nil, bundle: nil)
+       self.window?.makeKeyAndVisible()
+   }
+   
+   func setupRootViewController() {
+       
+       self.window?.rootViewController = UIHostingController(
+           rootView: ContactListView()
+       )
+   }
+   
+   func resetViewState(_ postResetBlock: @escaping () -> ()) {
+       if let rootViewController = self.window?.rootViewController {
+           if let _ = rootViewController.presentedViewController {
+               rootViewController.dismiss(animated: false, completion: postResetBlock)
+               return
+           }
+       }
+       postResetBlock()
+   }
+}
